+++ designer/client/cypress/e2e/fragment.cy.ts
@@ -217,7 +217,7 @@ describe("Fragment", () => {
             cy.get("[data-testid=window]")
                 .find("section")
                 .scrollTo("top")
-                .click() //drop focus from input - avoid random scrolling/reszing of window contents
+                .click(5, 5) //drop focus from input - avoid random scrolling/reszing of window contents
                 .matchImage();
 
             cy.get("[data-testid=window]").find("section").scrollTo("bottom").find('[id="nk-graph-fragment"]').should("be.visible");