+++ designer/client/src/components/graph/node-modal/customNode.tsx
@@ -1,4 +1,4 @@
-import type { PropsWithChildren} from "react";
+import type { PropsWithChildren } from "react";
 import React, { useMemo } from "react";
 
 import ProcessUtils from "../../../common/ProcessUtils";